requirements.txt
starlette==0.14.2
uvicorn[standard]
fastapi==0.65.1
psycopg2==2.8.6
pydantic==1.8.2
python-dotenv==0.17.1
python-jose==3.2.0
python-multipart==0.0.5
SQLAlchemy==1.4.15
